fix:struct-edit/files-adj
